tests/wpt/tests/html/browsers/history/the-location-interface/location-prototype-setting-goes-cross-origin-domain.sub.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>[[SetPrototypeOf]] on a Location object should not allow changing its value: cross-origin via document.domain after initially getting the object</title>
<link rel="help" href="http://html.spec.whatwg.org/multipage/#location-setprototypeof">
<link rel="author" title="Domenic Denicola" href="mailto:d@domenic.me">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/test-setting-immutable-prototype.js"></script>

<iframe src="/common/blank.html"></iframe>

<script>
"use strict";
setup({ explicit_done: true });

window.onload = () => {
  const targetLocation = frames[0].location;
  const origProto = Object.getPrototypeOf(targetLocation);

  test(() => {
    assert_not_equals(origProto, null);
  }, "Same-origin (for now): the prototype is accessible");

  document.domain = "{{host}}";

  test(() => {
    assert_equals(Object.getPrototypeOf(targetLocation), null);
  }, "Became cross-origin via document.domain: the prototype is now null");

  testSettingImmutablePrototype("Became cross-origin via document.domain", targetLocation, null, { isSameOriginDomain: false });

  testSettingImmutablePrototypeToNewValueOnly(
    "Became cross-origin via document.domain", targetLocation, origProto,
    "the original value from before going cross-origin", { isSameOriginDomain: false });

  done();
};
</script>
